+++ nautilus_core/model/src/events/mod.rs
@@ -14,6 +14,4 @@
 // -------------------------------------------------------------------------------------------------
 
 pub mod order;
-#[cfg(feature = "ffi")]
-pub mod order_api;
 pub mod position;
