+++ nautilus_core/model/src/python/mod.rs
@@ -17,9 +17,15 @@ use pyo3::{
     exceptions::PyValueError,
     prelude::*,
     types::{PyDict, PyList},
+    PyResult, Python,
 };
 use serde_json::Value;
 use strum::IntoEnumIterator;
+
+use crate::{data, enums, identifiers, instruments, orders};
+
+pub mod macros;
+pub mod types;
 
 pub const PY_MODULE_MODEL: &str = "nautilus_trader.core.nautilus_pyo3.model";
 
@@ -210,4 +216,79 @@ mod tests {
             );
         });
     }
+}
+
+/// Loaded as nautilus_pyo3.model
+#[pymodule]
+pub fn model(_: Python<'_>, m: &PyModule) -> PyResult<()> {
+    // data
+    m.add_class::<data::bar::BarSpecification>()?;
+    m.add_class::<data::bar::BarType>()?;
+    m.add_class::<data::bar::Bar>()?;
+    m.add_class::<data::order::BookOrder>()?;
+    m.add_class::<data::delta::OrderBookDelta>()?;
+    m.add_class::<data::quote::QuoteTick>()?;
+    m.add_class::<data::trade::TradeTick>()?;
+    // enums
+    m.add_class::<enums::AccountType>()?;
+    m.add_class::<enums::AggregationSource>()?;
+    m.add_class::<enums::AggressorSide>()?;
+    m.add_class::<enums::AssetClass>()?;
+    m.add_class::<enums::AssetType>()?;
+    m.add_class::<enums::BarAggregation>()?;
+    m.add_class::<enums::BookAction>()?;
+    m.add_class::<enums::BookType>()?;
+    m.add_class::<enums::ContingencyType>()?;
+    m.add_class::<enums::CurrencyType>()?;
+    m.add_class::<enums::InstrumentCloseType>()?;
+    m.add_class::<enums::LiquiditySide>()?;
+    m.add_class::<enums::MarketStatus>()?;
+    m.add_class::<enums::OmsType>()?;
+    m.add_class::<enums::OptionKind>()?;
+    m.add_class::<enums::OrderSide>()?;
+    m.add_class::<enums::OrderStatus>()?;
+    m.add_class::<enums::OrderType>()?;
+    m.add_class::<enums::PositionSide>()?;
+    m.add_class::<enums::PriceType>()?;
+    m.add_class::<enums::TimeInForce>()?;
+    m.add_class::<enums::TradingState>()?;
+    m.add_class::<enums::TrailingOffsetType>()?;
+    m.add_class::<enums::TriggerType>()?;
+    // identifiers
+    m.add_class::<identifiers::account_id::AccountId>()?;
+    m.add_class::<identifiers::client_id::ClientId>()?;
+    m.add_class::<identifiers::client_order_id::ClientOrderId>()?;
+    m.add_class::<identifiers::component_id::ComponentId>()?;
+    m.add_class::<identifiers::exec_algorithm_id::ExecAlgorithmId>()?;
+    m.add_class::<identifiers::instrument_id::InstrumentId>()?;
+    m.add_class::<identifiers::order_list_id::OrderListId>()?;
+    m.add_class::<identifiers::position_id::PositionId>()?;
+    m.add_class::<identifiers::strategy_id::StrategyId>()?;
+    m.add_class::<identifiers::symbol::Symbol>()?;
+    m.add_class::<identifiers::trade_id::TradeId>()?;
+    m.add_class::<identifiers::trader_id::TraderId>()?;
+    m.add_class::<identifiers::venue::Venue>()?;
+    m.add_class::<identifiers::venue_order_id::VenueOrderId>()?;
+    // orders
+    m.add_class::<orders::limit::LimitOrder>()?;
+    m.add_class::<orders::limit_if_touched::LimitIfTouchedOrder>()?;
+    m.add_class::<orders::market::MarketOrder>()?;
+    m.add_class::<orders::market_to_limit::MarketToLimitOrder>()?;
+    m.add_class::<orders::stop_limit::StopLimitOrder>()?;
+    m.add_class::<orders::stop_market::StopMarketOrder>()?;
+    m.add_class::<orders::trailing_stop_limit::TrailingStopLimitOrder>()?;
+    m.add_class::<orders::trailing_stop_market::TrailingStopMarketOrder>()?;
+    m.add_class::<crate::types::currency::Currency>()?;
+    m.add_class::<crate::types::money::Money>()?;
+    m.add_class::<crate::types::price::Price>()?;
+    m.add_class::<crate::types::quantity::Quantity>()?;
+    // instruments
+    m.add_class::<instruments::crypto_future::CryptoFuture>()?;
+    m.add_class::<instruments::crypto_perpetual::CryptoPerpetual>()?;
+    m.add_class::<instruments::currency_pair::CurrencyPair>()?;
+    m.add_class::<instruments::equity::Equity>()?;
+    m.add_class::<instruments::futures_contract::FuturesContract>()?;
+    m.add_class::<instruments::options_contract::OptionsContract>()?;
+    m.add_class::<instruments::synthetic::SyntheticInstrument>()?;
+    Ok(())
 }
