+++ nautilus_core/model/src/python/types/mod.rs
@@ -13,19 +13,7 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-pub mod currencies;
-pub mod data;
-pub mod enums;
-pub mod events;
-pub mod identifiers;
-pub mod instruments;
-pub mod macros;
-pub mod orderbook;
-pub mod orders;
-pub mod position;
-pub mod types;
-
-#[cfg(feature = "ffi")]
-pub mod ffi;
-#[cfg(feature = "python")]
-pub mod python;
+pub mod currency;
+pub mod money;
+pub mod price;
+pub mod quantity;
