+++ nautilus_core/model/src/types/currency.rs
@@ -14,23 +14,13 @@
 // -------------------------------------------------------------------------------------------------
 
 use std::{
-    ffi::{c_char, CStr},
     hash::{Hash, Hasher},
     str::FromStr,
 };
 
 use anyhow::{anyhow, Result};
-use nautilus_core::{
-    correctness::check_valid_string,
-    ffi::string::{cstr_to_string, str_to_cstr},
-    python::to_pyvalue_err,
-};
-use pyo3::{
-    exceptions::PyRuntimeError,
-    prelude::*,
-    pyclass::CompareOp,
-    types::{PyLong, PyString, PyTuple},
-};
+use nautilus_core::correctness::check_valid_string;
+use pyo3::prelude::*;
 use serde::{Deserialize, Serialize, Serializer};
 use ustr::Ustr;
 
@@ -153,245 +143,13 @@ impl<'de> Deserialize<'de> for Currency {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-// Python API
-////////////////////////////////////////////////////////////////////////////////
-#[cfg(feature = "python")]
-#[pymethods]
-impl Currency {
-    #[new]
-    fn py_new(
-        code: &str,
-        precision: u8,
-        iso4217: u16,
-        name: &str,
-        currency_type: CurrencyType,
-    ) -> PyResult<Self> {
-        Self::new(code, precision, iso4217, name, currency_type).map_err(to_pyvalue_err)
-    }
-
-    fn __setstate__(&mut self, py: Python, state: PyObject) -> PyResult<()> {
-        let tuple: (&PyString, &PyLong, &PyLong, &PyString, &PyString) = state.extract(py)?;
-        self.code = Ustr::from(tuple.0.extract()?);
-        self.precision = tuple.1.extract::<u8>()?;
-        self.iso4217 = tuple.2.extract::<u16>()?;
-        self.name = Ustr::from(tuple.3.extract()?);
-        self.currency_type = CurrencyType::from_str(tuple.4.extract()?).map_err(to_pyvalue_err)?;
-        Ok(())
-    }
-
-    fn __getstate__(&self, py: Python) -> PyResult<PyObject> {
-        Ok((
-            self.code.to_string(),
-            self.precision,
-            self.iso4217,
-            self.name.to_string(),
-            self.currency_type.to_string(),
-        )
-            .to_object(py))
-    }
-
-    fn __reduce__(&self, py: Python) -> PyResult<PyObject> {
-        let safe_constructor = py.get_type::<Self>().getattr("_safe_constructor")?;
-        let state = self.__getstate__(py)?;
-        Ok((safe_constructor, PyTuple::empty(py), state).to_object(py))
-    }
-
-    #[staticmethod]
-    fn _safe_constructor() -> PyResult<Self> {
-        Ok(Currency::AUD()) // Safe default
-    }
-
-    fn __richcmp__(&self, other: &Self, op: CompareOp, py: Python<'_>) -> Py<PyAny> {
-        match op {
-            CompareOp::Eq => self.eq(other).into_py(py),
-            CompareOp::Ne => self.ne(other).into_py(py),
-            _ => py.NotImplemented(),
-        }
-    }
-
-    fn __hash__(&self) -> isize {
-        self.code.precomputed_hash() as isize
-    }
-
-    fn __str__(&self) -> &'static str {
-        self.code.as_str()
-    }
-
-    fn __repr__(&self) -> String {
-        format!("{:?}", self)
-    }
-
-    #[getter]
-    #[pyo3(name = "code")]
-    fn py_code(&self) -> &'static str {
-        self.code.as_str()
-    }
-
-    #[getter]
-    #[pyo3(name = "precision")]
-    fn py_precision(&self) -> u8 {
-        self.precision
-    }
-
-    #[getter]
-    #[pyo3(name = "iso4217")]
-    fn py_iso4217(&self) -> u16 {
-        self.iso4217
-    }
-
-    #[getter]
-    #[pyo3(name = "name")]
-    fn py_name(&self) -> &'static str {
-        self.name.as_str()
-    }
-
-    #[getter]
-    #[pyo3(name = "currency_type")]
-    fn py_currency_type(&self) -> CurrencyType {
-        self.currency_type
-    }
-
-    #[staticmethod]
-    #[pyo3(name = "is_fiat")]
-    fn py_is_fiat(code: &str) -> PyResult<bool> {
-        Currency::is_fiat(code).map_err(to_pyvalue_err)
-    }
-
-    #[staticmethod]
-    #[pyo3(name = "is_crypto")]
-    fn py_is_crypto(code: &str) -> PyResult<bool> {
-        Currency::is_crypto(code).map_err(to_pyvalue_err)
-    }
-
-    #[staticmethod]
-    #[pyo3(name = "is_commodity_backed")]
-    fn py_is_commodidity_backed(code: &str) -> PyResult<bool> {
-        Currency::is_commodity_backed(code).map_err(to_pyvalue_err)
-    }
-
-    #[staticmethod]
-    #[pyo3(name = "from_str")]
-    #[pyo3(signature = (value, strict = false))]
-    fn py_from_str(value: &str, strict: bool) -> PyResult<Currency> {
-        match Currency::from_str(value) {
-            Ok(currency) => Ok(currency),
-            Err(e) => {
-                if strict {
-                    Err(to_pyvalue_err(e))
-                } else {
-                    // SAFETY: Safe default arguments for the unwrap
-                    let new_crypto =
-                        Currency::new(value, 8, 0, value, CurrencyType::Crypto).unwrap();
-                    Ok(new_crypto)
-                }
-            }
-        }
-    }
-
-    #[staticmethod]
-    #[pyo3(name = "register")]
-    #[pyo3(signature = (currency, overwrite = false))]
-    fn py_register(currency: Currency, overwrite: bool) -> PyResult<()> {
-        Currency::register(currency, overwrite).map_err(|e| PyRuntimeError::new_err(e.to_string()))
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// C API
-////////////////////////////////////////////////////////////////////////////////
-/// Returns a [`Currency`] from pointers and primitives.
-///
-/// # Safety
-///
-/// - Assumes `code_ptr` is a valid C string pointer.
-/// - Assumes `name_ptr` is a valid C string pointer.
-#[cfg(feature = "ffi")]
-#[no_mangle]
-pub unsafe extern "C" fn currency_from_py(
-    code_ptr: *const c_char,
-    precision: u8,
-    iso4217: u16,
-    name_ptr: *const c_char,
-    currency_type: CurrencyType,
-) -> Currency {
-    assert!(!code_ptr.is_null(), "`code_ptr` was NULL");
-    assert!(!name_ptr.is_null(), "`name_ptr` was NULL");
-
-    Currency::new(
-        CStr::from_ptr(code_ptr)
-            .to_str()
-            .expect("CStr::from_ptr failed for `code_ptr`"),
-        precision,
-        iso4217,
-        CStr::from_ptr(name_ptr)
-            .to_str()
-            .expect("CStr::from_ptr failed for `name_ptr`"),
-        currency_type,
-    )
-    .unwrap()
-}
-
-#[no_mangle]
-pub extern "C" fn currency_to_cstr(currency: &Currency) -> *const c_char {
-    str_to_cstr(format!("{currency:?}").as_str())
-}
-
-#[no_mangle]
-pub extern "C" fn currency_code_to_cstr(currency: &Currency) -> *const c_char {
-    str_to_cstr(&currency.code)
-}
-
-#[no_mangle]
-pub extern "C" fn currency_name_to_cstr(currency: &Currency) -> *const c_char {
-    str_to_cstr(&currency.name)
-}
-
-#[no_mangle]
-pub extern "C" fn currency_hash(currency: &Currency) -> u64 {
-    currency.code.precomputed_hash()
-}
-
-#[no_mangle]
-pub extern "C" fn currency_register(currency: Currency) {
-    CURRENCY_MAP
-        .lock()
-        .unwrap()
-        .insert(currency.code.to_string(), currency);
-}
-
-/// # Safety
-///
-/// - Assumes `code_ptr` is borrowed from a valid Python UTF-8 `str`.
-#[no_mangle]
-pub unsafe extern "C" fn currency_exists(code_ptr: *const c_char) -> u8 {
-    let code = cstr_to_string(code_ptr);
-    u8::from(CURRENCY_MAP.lock().unwrap().contains_key(&code))
-}
-
-/// # Safety
-///
-/// - Assumes `code_ptr` is borrowed from a valid Python UTF-8 `str`.
-#[no_mangle]
-pub unsafe extern "C" fn currency_from_cstr(code_ptr: *const c_char) -> Currency {
-    let code = cstr_to_string(code_ptr);
-    Currency::from_str(&code).unwrap()
-}
-
-////////////////////////////////////////////////////////////////////////////////
 // Tests
 ////////////////////////////////////////////////////////////////////////////////
 #[cfg(test)]
 mod tests {
-    use std::ffi::{CStr, CString};
-
-    use nautilus_core::ffi::string::str_to_cstr;
     use rstest::rstest;
 
-    use super::*;
-    use crate::{
-        enums::CurrencyType,
-        types::currency::{currency_exists, Currency},
-    };
+    use crate::{enums::CurrencyType, types::currency::Currency};
 
     #[rstest]
     #[should_panic(expected = "`Currency` code")]
@@ -444,80 +202,5 @@ mod tests {
         let serialized = serde_json::to_string(&currency).unwrap();
         let deserialized: Currency = serde_json::from_str(&serialized).unwrap();
         assert_eq!(currency, deserialized);
-    }
-
-    #[rstest]
-    fn test_registration() {
-        let currency = Currency::new("MYC", 4, 0, "My Currency", CurrencyType::Crypto).unwrap();
-        currency_register(currency);
-        unsafe {
-            assert_eq!(currency_exists(str_to_cstr("MYC")), 1);
-        }
-    }
-
-    #[rstest]
-    fn test_currency_from_py() {
-        let code = CString::new("MYC").unwrap();
-        let name = CString::new("My Currency").unwrap();
-        let currency = unsafe {
-            super::currency_from_py(code.as_ptr(), 4, 0, name.as_ptr(), CurrencyType::Crypto)
-        };
-        assert_eq!(currency.code.as_str(), "MYC");
-        assert_eq!(currency.name.as_str(), "My Currency");
-        assert_eq!(currency.currency_type, CurrencyType::Crypto);
-    }
-
-    #[rstest]
-    fn test_currency_to_cstr() {
-        let currency = Currency::USD();
-        let cstr = unsafe { CStr::from_ptr(currency_to_cstr(&currency)) };
-        let expected_output = format!("{:?}", currency);
-        assert_eq!(cstr.to_str().unwrap(), expected_output);
-    }
-
-    #[rstest]
-    fn test_currency_code_to_cstr() {
-        let currency = Currency::USD();
-        let cstr = unsafe { CStr::from_ptr(currency_code_to_cstr(&currency)) };
-        assert_eq!(cstr.to_str().unwrap(), "USD");
-    }
-
-    #[rstest]
-    fn test_currency_name_to_cstr() {
-        let currency = Currency::USD();
-        let cstr = unsafe { CStr::from_ptr(currency_name_to_cstr(&currency)) };
-        assert_eq!(cstr.to_str().unwrap(), "United States dollar");
-    }
-
-    #[rstest]
-    fn test_currency_hash() {
-        let currency = Currency::USD();
-        let hash = super::currency_hash(&currency);
-        assert_eq!(hash, currency.code.precomputed_hash());
-    }
-
-    #[rstest]
-    fn test_currency_from_cstr() {
-        let code = CString::new("USD").unwrap();
-        let currency = unsafe { currency_from_cstr(code.as_ptr()) };
-        assert_eq!(currency, Currency::USD());
-    }
-
-    #[rstest]
-    #[should_panic(expected = "`code_ptr` was NULL")]
-    fn test_currency_from_py_null_code_ptr() {
-        let name = CString::new("My Currency").unwrap();
-        let _ = unsafe {
-            currency_from_py(std::ptr::null(), 4, 0, name.as_ptr(), CurrencyType::Crypto)
-        };
-    }
-
-    #[rstest]
-    #[should_panic(expected = "`name_ptr` was NULL")]
-    fn test_currency_from_py_null_name_ptr() {
-        let code = CString::new("MYC").unwrap();
-        let _ = unsafe {
-            currency_from_py(code.as_ptr(), 4, 0, std::ptr::null(), CurrencyType::Crypto)
-        };
     }
 }
