+++ nautilus_core/model/src/types/price.rs
@@ -15,7 +15,6 @@
 
 use std::{
     cmp::Ordering,
-    collections::hash_map::DefaultHasher,
     fmt::{Debug, Display, Formatter},
     hash::{Hash, Hasher},
     ops::{Add, AddAssign, Deref, Mul, Neg, Sub, SubAssign},
@@ -23,17 +22,9 @@ use std::{
 };
 
 use anyhow::Result;
-use nautilus_core::{
-    correctness::check_f64_in_range_inclusive,
-    parsing::precision_from_str,
-    python::{get_pytype_name, to_pytype_err, to_pyvalue_err},
-};
-use pyo3::{
-    prelude::*,
-    pyclass::CompareOp,
-    types::{PyFloat, PyLong, PyTuple},
-};
-use rust_decimal::{Decimal, RoundingStrategy};
+use nautilus_core::{correctness::check_f64_in_range_inclusive, parsing::precision_from_str};
+use pyo3::prelude::*;
+use rust_decimal::Decimal;
 use serde::{Deserialize, Deserializer, Serialize};
 use thousands::Separable;
 
@@ -295,394 +286,6 @@ impl<'de> Deserialize<'de> for Price {
         let price: Price = price_str.into();
         Ok(price)
     }
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Python API
-////////////////////////////////////////////////////////////////////////////////
-#[cfg(feature = "python")]
-#[pymethods]
-impl Price {
-    #[new]
-    fn py_new(value: f64, precision: u8) -> PyResult<Self> {
-        Price::new(value, precision).map_err(to_pyvalue_err)
-    }
-
-    fn __setstate__(&mut self, py: Python, state: PyObject) -> PyResult<()> {
-        let tuple: (&PyLong, &PyLong) = state.extract(py)?;
-        self.raw = tuple.0.extract()?;
-        self.precision = tuple.1.extract::<u8>()?;
-        Ok(())
-    }
-
-    fn __getstate__(&self, py: Python) -> PyResult<PyObject> {
-        Ok((self.raw, self.precision).to_object(py))
-    }
-
-    fn __reduce__(&self, py: Python) -> PyResult<PyObject> {
-        let safe_constructor = py.get_type::<Self>().getattr("_safe_constructor")?;
-        let state = self.__getstate__(py)?;
-        Ok((safe_constructor, PyTuple::empty(py), state).to_object(py))
-    }
-
-    #[staticmethod]
-    fn _safe_constructor() -> PyResult<Self> {
-        Ok(Price::zero(0)) // Safe default
-    }
-
-    fn __add__(&self, other: PyObject, py: Python) -> PyResult<PyObject> {
-        if other.as_ref(py).is_instance_of::<PyFloat>() {
-            let other_float: f64 = other.extract(py)?;
-            Ok((self.as_f64() + other_float).into_py(py))
-        } else if let Ok(other_price) = other.extract::<Price>(py) {
-            Ok((self.as_decimal() + other_price.as_decimal()).into_py(py))
-        } else if let Ok(other_dec) = other.extract::<Decimal>(py) {
-            Ok((self.as_decimal() + other_dec).into_py(py))
-        } else {
-            let pytype_name = get_pytype_name(&other, py)?;
-            Err(to_pytype_err(format!(
-                "Unsupported type for __add__, was `{pytype_name}`"
-            )))
-        }
-    }
-
-    fn __radd__(&self, other: PyObject, py: Python) -> PyResult<PyObject> {
-        if other.as_ref(py).is_instance_of::<PyFloat>() {
-            let other_float: f64 = other.extract(py)?;
-            Ok((other_float + self.as_f64()).into_py(py))
-        } else if let Ok(other_price) = other.extract::<Price>(py) {
-            Ok((other_price.as_decimal() + self.as_decimal()).into_py(py))
-        } else if let Ok(other_dec) = other.extract::<Decimal>(py) {
-            Ok((other_dec + self.as_decimal()).into_py(py))
-        } else {
-            let pytype_name = get_pytype_name(&other, py)?;
-            Err(to_pytype_err(format!(
-                "Unsupported type for __radd__, was `{pytype_name}`"
-            )))
-        }
-    }
-
-    fn __sub__(&self, other: PyObject, py: Python) -> PyResult<PyObject> {
-        if other.as_ref(py).is_instance_of::<PyFloat>() {
-            let other_float: f64 = other.extract(py)?;
-            Ok((self.as_f64() - other_float).into_py(py))
-        } else if let Ok(other_price) = other.extract::<Price>(py) {
-            Ok((self.as_decimal() - other_price.as_decimal()).into_py(py))
-        } else if let Ok(other_dec) = other.extract::<Decimal>(py) {
-            Ok((self.as_decimal() - other_dec).into_py(py))
-        } else {
-            let pytype_name = get_pytype_name(&other, py)?;
-            Err(to_pytype_err(format!(
-                "Unsupported type for __sub__, was `{pytype_name}`"
-            )))
-        }
-    }
-
-    fn __rsub__(&self, other: PyObject, py: Python) -> PyResult<PyObject> {
-        if other.as_ref(py).is_instance_of::<PyFloat>() {
-            let other_float: f64 = other.extract(py)?;
-            Ok((other_float - self.as_f64()).into_py(py))
-        } else if let Ok(other_price) = other.extract::<Price>(py) {
-            Ok((other_price.as_decimal() - self.as_decimal()).into_py(py))
-        } else if let Ok(other_dec) = other.extract::<Decimal>(py) {
-            Ok((other_dec - self.as_decimal()).into_py(py))
-        } else {
-            let pytype_name = get_pytype_name(&other, py)?;
-            Err(to_pytype_err(format!(
-                "Unsupported type for __rsub__, was `{pytype_name}`"
-            )))
-        }
-    }
-
-    fn __mul__(&self, other: PyObject, py: Python) -> PyResult<PyObject> {
-        if other.as_ref(py).is_instance_of::<PyFloat>() {
-            let other_float: f64 = other.extract(py)?;
-            Ok((self.as_f64() * other_float).into_py(py))
-        } else if let Ok(other_price) = other.extract::<Price>(py) {
-            Ok((self.as_decimal() * other_price.as_decimal()).into_py(py))
-        } else if let Ok(other_dec) = other.extract::<Decimal>(py) {
-            Ok((self.as_decimal() * other_dec).into_py(py))
-        } else {
-            let pytype_name = get_pytype_name(&other, py)?;
-            Err(to_pytype_err(format!(
-                "Unsupported type for __mul__, was `{pytype_name}`"
-            )))
-        }
-    }
-
-    fn __rmul__(&self, other: PyObject, py: Python) -> PyResult<PyObject> {
-        if other.as_ref(py).is_instance_of::<PyFloat>() {
-            let other_float: f64 = other.extract(py)?;
-            Ok((other_float * self.as_f64()).into_py(py))
-        } else if let Ok(other_price) = other.extract::<Price>(py) {
-            Ok((other_price.as_decimal() * self.as_decimal()).into_py(py))
-        } else if let Ok(other_dec) = other.extract::<Decimal>(py) {
-            Ok((other_dec * self.as_decimal()).into_py(py))
-        } else {
-            let pytype_name = get_pytype_name(&other, py)?;
-            Err(to_pytype_err(format!(
-                "Unsupported type for __rmul__, was `{pytype_name}`"
-            )))
-        }
-    }
-
-    fn __truediv__(&self, other: PyObject, py: Python) -> PyResult<PyObject> {
-        if other.as_ref(py).is_instance_of::<PyFloat>() {
-            let other_float: f64 = other.extract(py)?;
-            Ok((self.as_f64() / other_float).into_py(py))
-        } else if let Ok(other_price) = other.extract::<Price>(py) {
-            Ok((self.as_decimal() / other_price.as_decimal()).into_py(py))
-        } else if let Ok(other_dec) = other.extract::<Decimal>(py) {
-            Ok((self.as_decimal() / other_dec).into_py(py))
-        } else {
-            let pytype_name = get_pytype_name(&other, py)?;
-            Err(to_pytype_err(format!(
-                "Unsupported type for __truediv__, was `{pytype_name}`"
-            )))
-        }
-    }
-
-    fn __rtruediv__(&self, other: PyObject, py: Python) -> PyResult<PyObject> {
-        if other.as_ref(py).is_instance_of::<PyFloat>() {
-            let other_float: f64 = other.extract(py)?;
-            Ok((other_float / self.as_f64()).into_py(py))
-        } else if let Ok(other_price) = other.extract::<Price>(py) {
-            Ok((other_price.as_decimal() / self.as_decimal()).into_py(py))
-        } else if let Ok(other_dec) = other.extract::<Decimal>(py) {
-            Ok((other_dec / self.as_decimal()).into_py(py))
-        } else {
-            let pytype_name = get_pytype_name(&other, py)?;
-            Err(to_pytype_err(format!(
-                "Unsupported type for __rtruediv__, was `{pytype_name}`"
-            )))
-        }
-    }
-
-    fn __floordiv__(&self, other: PyObject, py: Python) -> PyResult<PyObject> {
-        if other.as_ref(py).is_instance_of::<PyFloat>() {
-            let other_float: f64 = other.extract(py)?;
-            Ok((self.as_f64() / other_float).floor().into_py(py))
-        } else if let Ok(other_price) = other.extract::<Price>(py) {
-            Ok((self.as_decimal() / other_price.as_decimal())
-                .floor()
-                .into_py(py))
-        } else if let Ok(other_dec) = other.extract::<Decimal>(py) {
-            Ok((self.as_decimal() / other_dec).floor().into_py(py))
-        } else {
-            let pytype_name = get_pytype_name(&other, py)?;
-            Err(to_pytype_err(format!(
-                "Unsupported type for __floordiv__, was `{pytype_name}`"
-            )))
-        }
-    }
-
-    fn __rfloordiv__(&self, other: PyObject, py: Python) -> PyResult<PyObject> {
-        if other.as_ref(py).is_instance_of::<PyFloat>() {
-            let other_float: f64 = other.extract(py)?;
-            Ok((other_float / self.as_f64()).floor().into_py(py))
-        } else if let Ok(other_price) = other.extract::<Price>(py) {
-            Ok((other_price.as_decimal() / self.as_decimal())
-                .floor()
-                .into_py(py))
-        } else if let Ok(other_dec) = other.extract::<Decimal>(py) {
-            Ok((other_dec / self.as_decimal()).floor().into_py(py))
-        } else {
-            let pytype_name = get_pytype_name(&other, py)?;
-            Err(to_pytype_err(format!(
-                "Unsupported type for __rfloordiv__, was `{pytype_name}`"
-            )))
-        }
-    }
-
-    fn __mod__(&self, other: PyObject, py: Python) -> PyResult<PyObject> {
-        if other.as_ref(py).is_instance_of::<PyFloat>() {
-            let other_float: f64 = other.extract(py)?;
-            Ok((self.as_f64() % other_float).into_py(py))
-        } else if let Ok(other_price) = other.extract::<Price>(py) {
-            Ok((self.as_decimal() % other_price.as_decimal()).into_py(py))
-        } else if let Ok(other_dec) = other.extract::<Decimal>(py) {
-            Ok((self.as_decimal() % other_dec).into_py(py))
-        } else {
-            let pytype_name = get_pytype_name(&other, py)?;
-            Err(to_pytype_err(format!(
-                "Unsupported type for __mod__, was `{pytype_name}`"
-            )))
-        }
-    }
-
-    fn __rmod__(&self, other: PyObject, py: Python) -> PyResult<PyObject> {
-        if other.as_ref(py).is_instance_of::<PyFloat>() {
-            let other_float: f64 = other.extract(py)?;
-            Ok((other_float % self.as_f64()).into_py(py))
-        } else if let Ok(other_price) = other.extract::<Price>(py) {
-            Ok((other_price.as_decimal() % self.as_decimal()).into_py(py))
-        } else if let Ok(other_dec) = other.extract::<Decimal>(py) {
-            Ok((other_dec % self.as_decimal()).into_py(py))
-        } else {
-            let pytype_name = get_pytype_name(&other, py)?;
-            Err(to_pytype_err(format!(
-                "Unsupported type for __rmod__, was `{pytype_name}`"
-            )))
-        }
-    }
-    fn __neg__(&self) -> Decimal {
-        self.as_decimal().neg()
-    }
-
-    fn __pos__(&self) -> Decimal {
-        let mut value = self.as_decimal();
-        value.set_sign_positive(true);
-        value
-    }
-
-    fn __abs__(&self) -> Decimal {
-        self.as_decimal().abs()
-    }
-
-    fn __int__(&self) -> i64 {
-        self.as_f64() as i64
-    }
-
-    fn __float__(&self) -> f64 {
-        self.as_f64()
-    }
-
-    fn __round__(&self, ndigits: Option<u32>) -> Decimal {
-        self.as_decimal()
-            .round_dp_with_strategy(ndigits.unwrap_or(0), RoundingStrategy::MidpointNearestEven)
-    }
-
-    fn __richcmp__(&self, other: PyObject, op: CompareOp, py: Python<'_>) -> Py<PyAny> {
-        if let Ok(other_price) = other.extract::<Price>(py) {
-            match op {
-                CompareOp::Eq => self.eq(&other_price).into_py(py),
-                CompareOp::Ne => self.ne(&other_price).into_py(py),
-                CompareOp::Ge => self.ge(&other_price).into_py(py),
-                CompareOp::Gt => self.gt(&other_price).into_py(py),
-                CompareOp::Le => self.le(&other_price).into_py(py),
-                CompareOp::Lt => self.lt(&other_price).into_py(py),
-            }
-        } else if let Ok(other_dec) = other.extract::<Decimal>(py) {
-            match op {
-                CompareOp::Eq => (self.as_decimal() == other_dec).into_py(py),
-                CompareOp::Ne => (self.as_decimal() != other_dec).into_py(py),
-                CompareOp::Ge => (self.as_decimal() >= other_dec).into_py(py),
-                CompareOp::Gt => (self.as_decimal() > other_dec).into_py(py),
-                CompareOp::Le => (self.as_decimal() <= other_dec).into_py(py),
-                CompareOp::Lt => (self.as_decimal() < other_dec).into_py(py),
-            }
-        } else {
-            py.NotImplemented()
-        }
-    }
-
-    fn __hash__(&self) -> isize {
-        let mut h = DefaultHasher::new();
-        self.hash(&mut h);
-        h.finish() as isize
-    }
-
-    fn __str__(&self) -> String {
-        self.to_string()
-    }
-
-    fn __repr__(&self) -> String {
-        format!("Price('{self:?}')")
-    }
-
-    #[getter]
-    fn raw(&self) -> i64 {
-        self.raw
-    }
-
-    #[getter]
-    fn precision(&self) -> u8 {
-        self.precision
-    }
-
-    #[staticmethod]
-    #[pyo3(name = "from_raw")]
-    fn py_from_raw(raw: i64, precision: u8) -> PyResult<Price> {
-        Price::from_raw(raw, precision).map_err(to_pyvalue_err)
-    }
-
-    #[staticmethod]
-    #[pyo3(name = "zero")]
-    #[pyo3(signature = (precision = 0))]
-    fn py_zero(precision: u8) -> PyResult<Price> {
-        Price::new(0.0, precision).map_err(to_pyvalue_err)
-    }
-
-    #[staticmethod]
-    #[pyo3(name = "from_int")]
-    fn py_from_int(value: u64) -> PyResult<Price> {
-        Price::new(value as f64, 0).map_err(to_pyvalue_err)
-    }
-
-    #[staticmethod]
-    #[pyo3(name = "from_str")]
-    fn py_from_str(value: &str) -> PyResult<Price> {
-        Price::from_str(value).map_err(to_pyvalue_err)
-    }
-
-    #[pyo3(name = "is_zero")]
-    fn py_is_zero(&self) -> bool {
-        self.is_zero()
-    }
-
-    #[pyo3(name = "is_positive")]
-    fn py_is_positive(&self) -> bool {
-        self.is_positive()
-    }
-
-    #[pyo3(name = "as_double")]
-    fn py_as_double(&self) -> f64 {
-        fixed_i64_to_f64(self.raw)
-    }
-
-    #[pyo3(name = "as_decimal")]
-    fn py_as_decimal(&self) -> Decimal {
-        self.as_decimal()
-    }
-
-    #[pyo3(name = "to_formatted_str")]
-    fn py_to_formatted_str(&self) -> String {
-        self.to_formatted_string()
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// C API
-////////////////////////////////////////////////////////////////////////////////
-#[cfg(feature = "ffi")]
-#[no_mangle]
-pub extern "C" fn price_new(value: f64, precision: u8) -> Price {
-    // TODO: Document panic
-    Price::new(value, precision).unwrap()
-}
-
-#[cfg(feature = "ffi")]
-#[no_mangle]
-pub extern "C" fn price_from_raw(raw: i64, precision: u8) -> Price {
-    Price::from_raw(raw, precision).unwrap()
-}
-
-#[cfg(feature = "ffi")]
-#[no_mangle]
-pub extern "C" fn price_as_f64(price: &Price) -> f64 {
-    price.as_f64()
-}
-
-#[cfg(feature = "ffi")]
-#[no_mangle]
-pub extern "C" fn price_add_assign(mut a: Price, b: Price) {
-    a.add_assign(b);
-}
-
-#[cfg(feature = "ffi")]
-#[no_mangle]
-pub extern "C" fn price_sub_assign(mut a: Price, b: Price) {
-    a.sub_assign(b);
 }
 
 ////////////////////////////////////////////////////////////////////////////////
